# Conflicts:
#	packages/api-dynamodb-to-elasticsearch/src/index.ts
#	packages/api-file-manager-s3/src/plugins/graphqlFileStorageS3.ts
#	packages/api-file-manager/__tests__/files.test.ts
#	packages/api-file-manager/src/graphql/index.ts
#	packages/api-tenancy/src/createTenancy/createTenantsMethods.ts
#	packages/app-admin/src/components/FileManager/FileManagerView.tsx

+++ packages/api-file-manager-s3/src/plugins/graphqlFileStorageS3.ts
@@ -14,14 +14,14 @@ const plugin: GraphQLSchemaPlugin<FileManagerContext> = {
             input PreSignedPostPayloadInput {
                 name: String!
                 type: String!
-                size: Int!
+                size: Long!
             }
 
             type GetPreSignedPostPayloadResponseDataFile {
                 id: ID!
                 name: String!
                 type: String!
-                size: Int!
+                size: Long!
                 key: String!
             }
